chore(coordinator): Move websocket handling into own module
