+++ coordinator/src/orderbook/mod.rs
@@ -1,6 +1,7 @@
 pub mod db;
 pub mod routes;
 pub mod trading;
+pub mod websocket;
 
 #[cfg(test)]
 mod tests;
